Channel-bot - Kopie/requirements.txt
discord.py>=2.4.0
python-dotenv>=1.0.1
requests>=2.32.3
pillow>=10.4.0
pytesseract>=0.3.10
watchdog>=4.0.1
flask>=3.0.0
flask-cors>=4.0.0
aiohttp>=3.9.0
